[Update] R/W of np.array and MRF construction.
* To reduce file size, SIFT descriptors and labels of numpy objects are saved as int16 and int8, respectively.
* Update functions in order to apply any size images.
* Add comments in README.

+++ visual_word.py
@@ -12,7 +12,7 @@ import settings as st
 def load_blob_wise_sift(files):
     all_feature_list = []
     for f in files:
-        all_feature_list.append(np.load(f))
+        all_feature_list.append(np.load(f).astype(np.float64))
     return all_feature_list
